fix: email signals
